requirements.txt
# DSTARK Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=9.0.0

# Training and evaluation
tqdm>=4.62.0
pyyaml>=6.0
tensorboard>=2.10.0

# Data processing
scipy>=1.7.0
matplotlib>=3.4.0

# Optional for visualization
seaborn>=0.11.0
